platform-api-python/pyproject.toml
[project]
name = "platform-api-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "fastapi[standard]>=0.116.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.1",
    "pdfplumber>=0.11.7",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "tiktoken>=0.9.0",
    "unstructured[local-inference]>=0.18.11",
    "uvicorn>=0.35.0",
]
